+++ src/main/java/com/leonardo/tde/repository/PagamentoRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.leonardo.tde.domain.Pagamento;
 
 @Repository
-public interface PagamentoRepository extends JpaRepository<Pagamento, Long> {
+public interface PagamentoRepository extends JpaRepository<Pagamento, Integer> {
 }
